requirements.txt
PyQt6>=6.6.0
python-registry>=1.3.0
python-evtx>=0.7.4
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.1.0
Jinja2>=3.1.2
